some work on making the JEI integration for the crackhammer/ICrusherRecipe stuff work again - seems the animation isn't working, sadly
