+++ gradle.properties
@@ -3,7 +3,7 @@ top_mc_version=1.1x
 ccl_version=2.7.0.+
 ccc_version=2.4.1.+
 nei_version=2.3.0.+
-jei_version=4.4.1.+
+jei_version=4.5.0.+
 mantle_version=1.2.0.+
 tconstruct_version=2.7.0.+
 mcmp_version=2.0.1
